SW/powerdeck/requirements.txt
pyserial==3.5
pystray==0.19.4
PyAutoGUI==0.9.53